Simplify header analysis/reasoning by splitting core_io.h into the expected core_read.h/core_write.h (matching core_read.cpp/core_write.cpp)
