+++ src/rest.cpp
@@ -6,7 +6,8 @@
 #include <attributes.h>
 #include <chain.h>
 #include <chainparams.h>
-#include <core_io.h>
+#include <core_read.h>
+#include <core_write.h>
 #include <httpserver.h>
 #include <index/txindex.h>
 #include <primitives/block.h>
